Refactor: Inherit from NcclTestReportGenerationStrategy, remove redundant logic
