+++ src/cloudai/workloads/nccl_test/performance_report_generation_strategy.py
@@ -14,36 +14,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import re
 
-from cloudai import ReportGenerationStrategy, System, TestRun
+from cloudai import System, TestRun
 
 from .performance_report_generator import NcclTestPerformanceReportGenerator
+from .report_generation_strategy import NcclTestReportGenerationStrategy
 
 
-class NcclTestPerformanceReportGenerationStrategy(ReportGenerationStrategy):
+class NcclTestPerformanceReportGenerationStrategy(NcclTestReportGenerationStrategy):
     """Strategy for generating performance reports from NCCL test outputs."""
 
     def __init__(self, system: System, tr: TestRun) -> None:
         super().__init__(system, tr)
         self.performance_report = NcclTestPerformanceReportGenerator(self.test_run)
-
-    def can_handle_directory(self) -> bool:
-        stdout_path = self.test_run.output_path / "stdout.txt"
-        if stdout_path.exists():
-            with stdout_path.open("r") as file:
-                content = file.read()
-                return bool(
-                    re.search(r"out-of-place|in-place", content)
-                    and re.search(
-                        r"\b(size\s+count\s+type\s+redop\s+root\s+"
-                        r"time\s+algbw\s+busbw\s+#wrong\s+time\s+"
-                        r"algbw\s+busbw\s+#wrong)\b",
-                        content,
-                        re.IGNORECASE,
-                    )
-                )
-        return False
 
     def generate_report(self) -> None:
         self.performance_report.generate()
