+++ src/cloudai/workloads/nccl_test/prediction_report_generation_strategy.py
@@ -14,14 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import re
 
-from cloudai import ReportGenerationStrategy, System, TestRun
+from cloudai import System, TestRun
 
 from .prediction_report_generator import NcclTestPredictionReportGenerator
+from .report_generation_strategy import NcclTestReportGenerationStrategy
 
 
-class NcclTestPredictionReportGenerationStrategy(ReportGenerationStrategy):
+class NcclTestPredictionReportGenerationStrategy(NcclTestReportGenerationStrategy):
     """Strategy for generating prediction reports from NCCL test outputs."""
 
     def __init__(self, system: System, tr: TestRun) -> None:
@@ -32,23 +32,6 @@ class NcclTestPredictionReportGenerationStrategy(ReportGenerationStrategy):
 
     def _normalize_collective_type(self, subtest_name: str) -> str:
         return subtest_name.replace("_perf", "").replace("_mpi", "")
-
-    def can_handle_directory(self) -> bool:
-        stdout_path = self.test_run.output_path / "stdout.txt"
-        if stdout_path.exists():
-            with stdout_path.open("r") as file:
-                content = file.read()
-                return bool(
-                    re.search(r"out-of-place|in-place", content)
-                    and re.search(
-                        r"\b(size\s+count\s+type\s+redop\s+root\s+"
-                        r"time\s+algbw\s+busbw\s+#wrong\s+time\s+"
-                        r"algbw\s+busbw\s+#wrong)\b",
-                        content,
-                        re.IGNORECASE,
-                    )
-                )
-        return False
 
     def generate_report(self) -> None:
         self.prediction_report.generate()
